[MIG] website_sale_b2x_alt_price: Migration to 17.0
