+++ website_sale_b2x_alt_price/README.rst
@@ -61,7 +61,8 @@ Configuration
 
 To choose if you want to display or not the alternative price in the
 website products list, just enable or disable the feature from the
-"Customize" menu. This can be done independently per website:
+"Customize" tab of the web editor. This can be done independently per
+website:
 
 |toggle-list|
 
@@ -118,6 +119,7 @@ Contributors
 
    -  Jairo Llopis
    -  Carlos Roca
+   -  Carlos López
 
 Maintainers
 -----------
