+++ website_sale_b2x_alt_price/readme/CONTRIBUTORS.md
@@ -1,3 +1,4 @@
 - [Tecnativa](https://www.tecnativa.com/):
   - Jairo Llopis
   - Carlos Roca
+  - Carlos López
