+++ website_sale_b2x_alt_price/static/description/index.html
@@ -401,7 +401,8 @@ detail page:</p>
 <h1><a class="toc-backref" href="#toc-entry-1">Configuration</a></h1>
 <p>To choose if you want to display or not the alternative price in the
 website products list, just enable or disable the feature from the
-“Customize” menu. This can be done independently per website:</p>
+“Customize” tab of the web editor. This can be done independently per
+website:</p>
 <p><img alt="toggle-list" src="https://raw.githubusercontent.com/OCA/e-commerce/17.0/website_sale_b2x_alt_price/static/description/toggle-list.gif" /></p>
 <p>You can also toggle the behavior similarly in the product details page.
 This can also be done independently per website:</p>
@@ -450,6 +451,7 @@ If you spotted it first, help us to smash it by providing a detailed and welcome
 <li><a class="reference external" href="https://www.tecnativa.com/">Tecnativa</a>:<ul>
 <li>Jairo Llopis</li>
 <li>Carlos Roca</li>
+<li>Carlos López</li>
 </ul>
 </li>
 </ul>
